potentially fix memory leak
Currently, a gc process runs on a timer to purportedly remove keys
associated with stale subscription browser clients from the redis store.
To do this, this plugin stores each client ID in a sorted set(ZSET)
using the last ping time of the client as the score. Thus if a client
does not ping the server for the timeout period at the /faye endpoint,
the client ID gets stale and, in theory, the gc process collects all the
keys associated with this client on the next gc run, which runs every
minute.
The plugin stores the subscriptions to channels for a client in keys of
the form: {shared}/channels<channel name>. Each of these keys stores a
set containing all the client IDs that subscribe to this channel. There
could be 1000s of these channels at any given point of time. Therefore
to make it easy to GC these subscriptions, the plugin stores all the
channels to which a client subscribes at a key of the form:
/clients/<client Id>/channels. In addition, when a message is published
to faye, the publish method on this plugin is invoked to ensure that
clients connected to other instances of the node app also get the
notifications. The publish method stores the message in a list with a
key of the form: /clients/<client Id>/messages. It then publishes the
client Id on a redis channel that all instances of this plugin
(via the node apps) are subscribed to. When each subscriber (node app)
receives this message, it checks if it has a connection to the
specified client ID. If it does, it delivers all the messages in the
list to the client and deletes the key.
Now, what seems to be happening is that somehow the client ID in the
sorted set (ZSET) seems to be garbage collected. However, due to some
race condition, the client ID stays in the
{shared}/channels<channel name> channel subscriptions. Therefore when
the publish method is called, it pushes the message to all the client
IDs associated with the channel (which includes this client ID which
has already been garbage collected). Since none of the subscribers now
have a connection to the browser with that client ID, this key never
gets collected, thus causing a memory leak.
The only way this could possibly happen is if somehow a subscription
call comes in just after the gc collects a client ID - say because the
browser sat idle for longer than the timeout period (without sending a
ping) and then subscribed to a channel. At this point, prior to this
patch, the client ID will be added to the subscription, but now will no
longer be present in the sorted set. However, this also requires that
the browser no longer send a ping with the same client ID. This is
somewhat inexplicable. However, this patch now also adds the client ID
to the sorted set on a subscription call to ensure that regardless of
whether the ping is called or not after this point, the newly added
subscription will be garbage collected eventually.